5th commit

+++ libft/ft_putnbr.c
@@ -6,13 +6,13 @@
 /*   By: rnomura <rnomura@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/18 20:16:10 by rnomura           #+#    #+#             */
-/*   Updated: 2024/05/18 23:20:34 by rnomura          ###   ########.fr       */
+/*   Updated: 2024/05/19 22:46:05 by rnomura          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_putnbr(unsigned long long n)
+int	ft_putnbr(unsigned int n)
 {
 	int count;
